:recycle: (Tutorial) extract generic creditTimeline from onboardingTimeline
